[Lighting] Darken the whole level and add lighting around all fire sources when active
Setup 2D render layers for everything including occlusion/shadow, although not used yet

We use Mix lighting to avoid stacking too much light when multiple sources are close.

Note that flames should use custom material to avoid being darkened, but for Torch the sprite is currently part of the prop body so we cannot do it. But it works because the Torch's flame lit by itself and Mix avoids overlighting.