[Fireball] Manually revert the important numerical changes in b9ba1c70 "Fixed sprite offset compared to collider"
This was a mistake in the first place, as the offset was intentional since fireball is hovering over ground so the collider was indeed projected on ground and lower than the sprite. Also reverted FireballSpawnPoint changes in Throw animations but kept the 4 -> 3 offset in the orthogonal direction. Kept other good modifications.

Since torch is higher than fireball you can still hit the torch unexpectedly when aiming low, but at least it covers more space than the player expects, not less.