[Light] Start with Fire Pit off more darker room, add additive Room Light to Fire Pit
Note that we don't have a Max Blending mode, and Mixing Energy < 1 (Room Light should be around 0.8 in Mix blending) with Energy >= 1 will cause the light to be darker due to average (if you try mixing lights both < 1 or > 1 then it triggers indeterministic bug https://github.com/godotengine/godot/issues/77889 which is even worse).

So additive light for the whole room is still the safest. Although because of the bug mentioned above and because even a Torch should show at least the original sprite color, we prefer keeping all lights Energy to 1 exactly which means final lighting will be more than the sprite original color, so not really what we want.